fix: update stale imports for moved metric plot classes
The metric plot classes (MetricByHorizonV2Mean, MetricByTimePeriodV2Mean,
MetricMapV2, RegionalMetricDistributionPlot, etc.) moved to
chap_core/assessment/metric_plots/ but several call sites still imported
them from chap_core.plotting.evaluation_plot, causing import-time
failures that blocked the entire test suite from collecting. Update
those call sites to import from the new package.

Also drop the duplicated CHAP_VERSION lookup in generate_modelcard.py
in favour of the existing chap_core.__version__.
